Pruebas Unitarias

+++ src/models/MovieResult.h
@@ -2,21 +2,25 @@
 #define MOVIE_RESULT_H
 
 #include <string>
-#include <nlohmann/json.hpp>
+#include <nlohmann/json.hpp>  // Biblioteca para manipular objetos JSON
 
+// Estructura que representa la información procesada de una película
+// después de combinar datos de detalles, créditos y géneros.
 struct MovieResult {
-    long tmdbId;
-    std::string title;
-    std::string year;
-    std::string imageUrl;
-    std::string description;
-    std::string genres;
-    std::string director;
-    std::string writers;
-    std::string actors;
-    int generalRating;
+    long tmdbId;              // ID de la película en The Movie Database (TMDb)
+    std::string title;        // Título de la película
+    std::string year;         // Año de estreno (extraído de release_date)
+    std::string imageUrl;     // URL completa del póster de la película
+    std::string description;  // Descripción o sinopsis
+    std::string genres;       // Lista de géneros separados por coma (ej. "Acción, Aventura")
+    std::string director;     // Nombre del director principal
+    std::string writers;      // Lista de guionistas separados por coma
+    std::string actors;       // Lista de actores principales separados por coma
+    double generalRating;     // Calificación promedio de la película (vote_average)
 };
 
+// Función para convertir un objeto MovieResult a JSON
+// Utilizada por la biblioteca nlohmann::json al serializar respuestas
 inline void to_json(nlohmann::json& j, const MovieResult& m) {
     j = {
         {"tmdbId", m.tmdbId},
@@ -33,4 +37,3 @@ inline void to_json(nlohmann::json& j, const MovieResult& m) {
 }
 
 #endif // MOVIE_RESULT_H
-
